GH-9 Switch report destination argument to File object
In version 4.0 of Gradle, the ConfigurableReport.setDestination method
was overloaded to accept either a generic Object or a File (previously,
only generic object was accepted). The original method was deprecated,
causing a warning to be printed when the merge coverage reports plug-in
is applied, as it configures the destination as a string.

This change switches to configuring the location as a File, which should
cause the newer method to be used when available