Resolves GH-20: Force correct report configuration for merge task
This change set adds a workaround for a Gradle configuration issue,
where some configuration done in the constructor of the created task
would not apply. Settings would read correctly within the constructor,
but once back in the plug-in configuration, would read incorrectly.

This change adds re-configuration of these values at the plug-in
configuration level, forcing the intended configuration