Resolves GH-28: Create a replacement plugin to remove task name conflict
The Gradle project-report plug-in added tasks with the same name as
tasks added by the dependency-insight plug-in. To avoid this, this
change set adds an alternate dependency-reporting plug-in which uses
non-conflicting names, and deprecates the dependency-insight plug-in.
The original plug-in was not updated to maintain compatibility
overlapping a major release, and may be removed in 3.x or later